+++ ios/Classes/query/OnArtistQuery.swift
@@ -19,11 +19,18 @@ class OnArtistsQuery {
         // the [Artist]. The others will be sorted "manually" using
         // [formatSongList] before send to Dart.
         
-        // Query everything in background for a better performance.
-        loadArtists(cursor: cursor.collections)
+        // We cannot "query" without permission so, just return a empty list.
+        let hasPermission = SwiftOnAudioQueryPlugin().checkPermission()
+        if hasPermission {
+            // Query everything in background for a better performance.
+            loadArtists(cursor: cursor.collections)
+        } else {
+            // There's no permission so, return empty to avoid crashes.
+            result([])
+        }
     }
     
-    internal func loadArtists(cursor: [MPMediaItemCollection]!) {
+    private func loadArtists(cursor: [MPMediaItemCollection]!) {
         DispatchQueue.global(qos: .userInitiated).async {
             var listOfArtists: [[String: Any?]] = Array()